Created LocalSettingsHelper.java
This helper class allows settings to be persisted locally by the instrument bootstrap mechanism. With this we can
store a list of parameters for an instrument interface running on a specific machine (ex: communication port), so
they can retrieved the next time the interface is launched.
  
Practically, this allows a user to setup an instrument interface the first time it is launched, without having to
worry about setting it up on each run, because the local configuration is kept on disk, and can be retrieved automatically.

Tbf310InstrumentRunner.java, instrument-context.xml: User can now select the communication port from a configuration menu.  The selected com port is persisted locally using the LocalSettingsHelper (see above).