ONYX-553
Values of the pdf consent form specified in the module-config.properties are added as variables when exporting to dcc.
Variables are of Datatype Text (signatures and images are not currently exported to variables. If needed, we will have to do it in a second time).
Modification of MarbleModule, module-context and module-config.properties