onyx-1008 onyx-1009 onyx-1015
Marble modifications for the consent language choice when mode manual only is available (bug in Alberta cohort fixed)
New variable for locale added in module-config.properties.
Methods for setting and getting this variable added in ConsentServiceImpl
Modifications in ConsentModeSelectionPanel to display the dropdown locale list when manual mode is the only available and in ManualConsentPanel to set the locale with the variable value if only one language is available on manual mode (in this case, the first consent step screen is not display as there is nothing to choose).